requirements.txt
flask==2.3.3
pytesseract==0.3.10
pillow==9.5.0
pandas==2.0.3
opencv-python==4.8.0.76
numpy==1.24.3
regex==2023.6.3
werkzeug==2.3.7